google_sheets_mcp/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
mcp
pandas
